+++ vk-api-methods/src/main/kotlin/name/anton3/vkapi/methods/longpoll/events/ChatEdited.kt
@@ -2,6 +2,6 @@ package name.anton3.vkapi.methods.longpoll.events
 
 data class ChatEdited(
     override val eventType: Int,
-    val chatId: Long,
+    val chatId: Int,
     val self: Boolean
 ) : LongPollEvent(eventType)
