+++ vk-api-methods/src/main/kotlin/name/anton3/vkapi/methods/longpoll/events/FriendOnline.kt
@@ -6,7 +6,7 @@ import name.anton3.vkapi.vktypes.VkDate
 
 data class FriendOnline(
     override val eventType: Int,
-    val userId: Long,
+    val userId: Int,
     val platform: FriendPlatform?,
     val timestamp: VkDate
 ) : LongPollEvent(eventType) {
@@ -14,11 +14,11 @@ data class FriendOnline(
     companion object {
         @JvmStatic
         @JsonCreator(mode = JsonCreator.Mode.PROPERTIES)
-        fun parse(eventType: Int, userId: Long, extra: Long, timestamp: VkDate): FriendOnline {
+        fun parse(eventType: Int, userId: Int, extra: Int, timestamp: VkDate): FriendOnline {
             return FriendOnline(
                 eventType,
                 -userId,
-                (extra and 0xFF).takeIf { it != 0L }?.let { FriendPlatform.parse(it) },
+                (extra and 0xFF).takeIf { it != 0 }?.let { FriendPlatform.parse(it) },
                 timestamp
             )
         }
